requirements.txt
prefect>=3.0.0
python-dotenv>=0.19.0
pandas>=2.2.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.32.0
pdfplumber>=0.10.3
openai>=1.12.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
plotly>=5.20.0